requirements.txt
Flask
Flask-SQLAlchemy
flask-restful
python-dotenv
Flask-Security
email_validator
bcrypt
flask-jwt-extended